refactor(child-pi): H-7 step 1 — extract transcript batching + compaction to child-pi-transcript.ts
Decomposed runChildPi's first 170 lines into a dedicated module. The
transcript batching (50ms debounced flush, O_NOFOLLOW | O_CREAT | O_APPEND)
and compaction helpers (compactString, compactValue) were module-level
functions with no closure dependencies on runChildPi's state — safe to
extract with zero behavior change.

child-pi.ts: 1842 → 1672 lines (-170). New child-pi-transcript.ts: 169 lines.
Public API preserved via re-exports so external consumers (background-runner,
dynamic-workflow-context, goal-evaluator, etc.) are unaffected.

All 65 child-pi tests + 10 dependent tests pass (no regressions).
